Enforce SAL architecture boundaries and clean up data access flow

+++ backend/storage_access_layer/db/schema.py
@@ -27,24 +27,24 @@ class ManifestMetrics(ManifestBase):
 
     avg_bbox_size: Mapped[int] = mapped_column(Integer, nullable=False)
     std_dev_bounding_box_area: Mapped[Float] = mapped_column(Float, nullable=True)
-    variance_bounding_box_area: Mapped[Float] = mapped_column(Float, nullable=True)
-    mean_width: Mapped[Float] = mapped_column(Float, nullable=True)
-    mean_height: Mapped[Float] = mapped_column(Float, nullable=True)
-    variance_bounding_box_width: Mapped[Float] = mapped_column(Float, nullable=True)
-    variance_bounding_box_height: Mapped[Float] = mapped_column(Float, nullable=True)
+    variance_bounding_box_area: Mapped[int] = mapped_column(Integer, nullable=True)
+    mean_width: Mapped[int] = mapped_column(Integer, nullable=True)
+    mean_height: Mapped[int] = mapped_column(Integer, nullable=True)
+    variance_bounding_box_width: Mapped[int] = mapped_column(Integer, nullable=True)
+    variance_bounding_box_height: Mapped[int] = mapped_column(Integer, nullable=True)
 
     step_count: Mapped[int] = mapped_column(Integer, nullable=False)
     step_count_on_path: Mapped[int] = mapped_column(Integer, nullable=True)
     total_trial_area: Mapped[Float] = mapped_column(Float, nullable=True)
     mean_step_distance: Mapped[Float] = mapped_column(Float, nullable=True)
-    variance_step_distance: Mapped[Float] = mapped_column(Float, nullable=True)
+    variance_step_distance: Mapped[int] = mapped_column(Integer, nullable=True)
     active_trial_duration_all: Mapped[Float] = mapped_column(Float, nullable=True)
     active_trial_duration_path: Mapped[Float] = mapped_column(Float, nullable=True)
     max_footstep_duration_frames: Mapped[int] = mapped_column(Integer, nullable=True)
 
-    mean_heading_angle: Mapped[Float] = mapped_column(Float, nullable=True)
-    std_heading_angle: Mapped[Float] = mapped_column(Float, nullable=True)
-    variance_heading_angle: Mapped[Float] = mapped_column(Float, nullable=True)
+    mean_heading_angle: Mapped[int] = mapped_column(Integer, nullable=True)
+    std_heading_angle: Mapped[int] = mapped_column(Integer, nullable=True)
+    variance_heading_angle: Mapped[int] = mapped_column(Integer, nullable=True)
 
 
 class ManifestFootstep(ManifestBase):
